Methods for LMLS

A couple of methods for location-scale regression models from the lmls() function are provided.

Usage

# S3 method for class 'lmls'
coef(object, predictor = c("location", "scale"), ...)

# S3 method for class 'lmls'
fitted(object, predictor = c("location", "scale"), ...)

# S3 method for class 'lmls'
predict(
  object,
  newdata = NULL,
  predictor = c("location", "scale"),
  type = c("link", "response"),
  ...
)

# S3 method for class 'lmls'
residuals(object, type = c("deviance", "pearson", "response"), ...)

# S3 method for class 'lmls'
vcov(object, predictor = c("location", "scale"), ...)

Arguments

object

A location-scale regression model from the lmls() function.

predictor

The predictor to work on. Either "location" or "scale" or both. If both, a list with the names "location" and "scale" is returned.

...

Currently ignored.

newdata

A data frame (or list or environment) with the covariate values at which the predictions are computed. If NULL, the predictions at the original data are returned.

type

Used by predict() and residuals():

For predict(), "link" or "response". If "link" (default), \(\mu\) and log(\(\sigma\)) are returned. If "response", \(\mu\) and \(\sigma\) are returned.

For residuals(), "deviance", "pearson" or "response". If "deviance" (default) or "pearson", (\(y - \mu\)) / \(\sigma\) is returned. If "response", \(y - \mu\) is returned.

Value

A numeric vector for residuals(). For the other methods, a numeric vector if the argument predictor is either "location" or "scale", or a list with the names location and scale if it is both.